spdx-utils: Remove the committed ScanCode license texts
The reporter can now read the ScanCode license texts from the install
location, so the committed license texts are not required anymore.

This ensures that the license texts are compatible with the installed
ScanCode version and reduces the size of the ORT distribution. On the
other hand it requires that ScanCode is installed to be able to resolve
the license texts.

Signed-off-by: Martin Nonnenmacher <martin.nonnenmacher@bosch.io>